# tsoa-test/build/swagger.yaml
components:
    examples: {}
    headers: {}
    parameters: {}
    requestBodies: {}
    responses: {}
    schemas:
        User:
            properties:
                id:
                    type: number
                    format: double
                email:
                    type: string
                name:
                    type: string
                status:
                    type: string
                    enum:
                        - Happy
                        - Sad
                phoneNumbers:
                    items:
                        type: string
                    type: array
            required:
                - id
                - email
                - name
                - phoneNumbers
            type: object
            additionalProperties: false
        Pick_User.email-or-name-or-phoneNumbers_:
            properties:
                email:
                    type: string
                name:
                    type: string
                phoneNumbers:
                    items:
                        type: string
                    type: array
            required:
                - email
                - name
                - phoneNumbers
            type: object
            description: 'From T, pick a set of properties whose keys are in the union K'
        UserCreationParams:
            $ref: '#/components/schemas/Pick_User.email-or-name-or-phoneNumbers_'
    securitySchemes: {}
info:
    title: tsoa-test
    version: 0.0.0
    contact: {}
openapi: 3.0.0
paths:
    '/tests/{testId}':
        get:
            operationId: GetTest
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                additionalProperties: false
                                type: object
            tags:
                - Tests
            security: []
            parameters:
                -
                    in: path
                    name: testId
                    required: true
                    schema:
                        format: double
                        type: number
    '/users/{userId}':
        get:
            operationId: GetUser
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/User'
            tags:
                - Users
            security: []
            parameters:
                -
                    in: path
                    name: userId
                    required: true
                    schema:
                        format: double
                        type: number
                -
                    in: query
                    name: name
                    required: false
                    schema:
                        type: string
    /users:
        post:
            operationId: CreateUser
            responses:
                '201':
                    description: Created
            tags:
                - Users
            security: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UserCreationParams'
servers:
    -
        url: /